[feat] Add author permission in modify, create, and delete

+++ myMovieReviews/review/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, redirect, get_object_or_404
+from django.contrib.auth.decorators import login_required
 from .models import Review
 from .forms import ReviewForm
 
@@ -19,31 +20,40 @@ def review_detail(request, pk):
     runningTimeModified = f"{review.runningTime//60}시간 {review.runningTime%60}분"
     return render(request, 'review/review_detail.html', {'review': review, 'runningTimeModified': runningTimeModified})
 
+@login_required(login_url='common:login')
 def review_create(request):
     if(request.method == 'POST'):
         form = ReviewForm(request.POST)
         if form.is_valid():
             review = form.save(commit=False)
+            review.author = request.user
             review.save()
             return redirect('review:index')
     else:
         form = ReviewForm()
     return render(request, 'review/review_form.html', {'form': form})
 
+@login_required(login_url='common:login')
 def review_modify(request, pk):
     review = get_object_or_404(Review, pk=pk)
-    if(request.method == 'POST'):
-        form = ReviewForm(request.POST, instance=review)
-        if form.is_valid():
-            review = form.save(commit=False)
-            review.save()
-            return redirect('review:review_detail', pk=pk)
+    if(request.user == review.author):
+        if(request.method == 'POST'):
+            form = ReviewForm(request.POST, instance=review)
+            if form.is_valid():
+                review = form.save(commit=False)
+                review.save()
+                return redirect('review:review_detail', pk=pk)
+        else:
+            form = ReviewForm(instance=review)
+        return render(request, 'review/review_form.html', {'form': form})
     else:
-        form = ReviewForm(instance=review)
-    return render(request, 'review/review_form.html', {'form': form})
+        return render(request, 'permission_denied.html')
 
+@login_required(login_url='common:login')
 def review_delete(request, pk):
-    if(request.method == 'POST'):
-        review = get_object_or_404(Review, pk=pk)
-        review.delete()
-    return redirect('review:index')
+    review = get_object_or_404(Review, pk=pk)
+    if (request.user == review.author):
+        if(request.method == 'POST'):     
+            review.delete()
+        return redirect('review:index')
+    return render(request, 'permission_denied.html')
